requirements.txt
pandas
plotly
numpy
streamlit_aggrid
pathlib
geopandas
folium
streamlit_folium
matplotlib
mapclassify



